+++ packages/atproto/CHANGELOG.md
@@ -9,6 +9,8 @@
   - `requestEmailConfirmation`
 - Supported `com.atproto.server.confirmEmail` on `ServersService`. ([#855](https://github.com/myConsciousness/atproto.dart/issues/855))
   - `requestEmailConfirmation`
+- Supported `com.atproto.server.updateEmail` on `ServersService`. ([#860](https://github.com/myConsciousness/atproto.dart/issues/860))
+  - `updateEmail`
 
 ## v0.6.8