+++ packages/bluesky/CHANGELOG.md
@@ -8,6 +8,8 @@
   - `requestEmailConfirmation`
 - Supported `com.atproto.server.confirmEmail` on `ServersService`. ([#855](https://github.com/myConsciousness/atproto.dart/issues/855))
   - `requestEmailConfirmation`
+- Supported `com.atproto.server.updateEmail` on `ServersService`. ([#860](https://github.com/myConsciousness/atproto.dart/issues/860))
+  - `updateEmail`
 
 ## v0.8.25